Merge IcoMoon glyphs into custom icon-font
It's not worth lugging around an extra typeface for only three icons. As
this is only a user-created subset of custom fonts (not the real IcoMoon
library) there's little point in keeping it. Users aren't likely to have
made use of the extra icons in the font, since they weren't documented.