Disambiguate "On Changes" setting
If the user has coloured icons disabled, the onChanges option doesn't do
anything. This can be better communicated to users through a visual cue:
fading the checkbox makes it clear one setting is needed for the other.

The description's also been clarified to be more informative for users;
it wasn't obvious the setting only affected paths under version control.