Prioritise common formats
This is admittedly microoptimising, but it's worth moving something like
".xml" to the top of a 120-item list of expressions (especially when the
majority are less-often used).

Nota bene: We normally try to avoid these types of ad-hoc optimisations,
as they make code less readable/traversable while yielding little gain.