Refactor config's handling of advanced properties
The previous array-based approach made sense when only three values were
being set for each match. With the introduction of alias and interpreter
properties, however, this approach has the opposite effect on legibility
that it sought to induce.

In an effort to keep the config both readable *and* maintainable, object
values are now used to specify the less-common attributes of each match:

 [".ext", "colour",  scope: "name", alias: "name", interpreter: "name"]

All three attributes may be set to the same value by passing a string as
the array's third element:

    [".ext", "colour", "name"]

Which would expand into:

    [".ext", "colour", scope: "name",
                       alias: "name",
                 interpreter: "name"]

This retains the conciseness of the original approach, but still remains
flexible and extendable. A key concern with this method is its behaviour
mightn't be obvious to all contributors - some users may even cargo-cult
the third parameter without understanding its purpose:

    # User adds a hypothetical language named "RadCool"
    [".rc", "medium-green", "rc"]

The blind inclusion of the third parameter would override the shell rule
when detecting hashbangs, showing the wrong icon for "#!/bin/rc".