Use null or false to identify shebang-match failure
For files with hashbangs and no recognised interpreter, having a default
icon in its stead is probably better than a default blank icon. We can't
use a null value to indicate this, since that's also returned when there
isn't a hashbang in a string.

Using two distinct types of falsy values is slightly unintuitive, but it
keeps us from splitting our logic even further.