Add icon-matching for popular frameworks/libraries
It's currently unclear if the "framework:" property will even be needed.
The plan is to implement a new package setting allowing users the choice
to restore the original JS/CSS icons. It's possible that some users will
find the sudden appearance of logos a little disconcerting.

It's far more likely I'm simply overthinking this, though.