Improve handling of serialised data
Instead of wiping outdated caches, an effort is now made to repair them.
The service will attempt to locate the correct icons using their classes
that were serialised from the last session. This approach is much slower
than simply referencing IconRules by their index, which is why we're not
using it unless necessary.