Scan headers when running file searches
Currently, filesystem scans are triggered from the tree-view, meaning an
ugly hack has to be used to update the Fuzzy-Finder's icons (since users
mightn't always have the tree-view open when running a search).

Obviously, the correct way of implementing this to begin with would have
been to queue *all* filesystem scans from iconClassForPath, collecting a
list of paths to send to a batch scan. This would obviate the need for a
tangled callback system that's overly-reliant on the tree-view's events.

A clean rewrite (preferably in pure JavaScript) can wait until v2.1.0 or
something. Hopefully this mess won't have to exist for too long.